// frontend/gen/calculator/v1/calculator_pb.ts
export enum Operation {
  /**
   * @generated from enum value: OPERATION_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: OPERATION_ADD = 1;
   */
  ADD = 1,

  /**
   * @generated from enum value: OPERATION_SUBTRACT = 2;
   */
  SUBTRACT = 2,

  /**
   * @generated from enum value: OPERATION_MULTIPLY = 3;
   */
  MULTIPLY = 3,

  /**
   * @generated from enum value: OPERATION_DIVIDE = 4;
   */
  DIVIDE = 4,
}